// frontend/ur-homie/src/types/forms.ts
export interface GeneralFormData {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  country: string;
  city: string;
  address: string;
}

export interface ProviderDetailsData {
  education: string;
  certifications: string;
  experienceDescription: string;
  startTime: string;
  endTime: string;
  coverageArea: string;
}